be/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# File handling
aiofiles==23.2.0
python-magic==0.4.27

# AI & LLM - FIXED VERSIONS TO AVOID CONFLICTS
openai>=1.6.1,<2.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
huggingface-hub==0.19.4
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.18
# qdrant-client==1.6.9  # Alternative vector DB

# OCR
easyocr==1.7.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Text processing
pypdf==3.17.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1

# Utilities
pydantic[email]==2.5.0  # Added [email] extra for EmailStr support
pydantic-settings==2.1.0
httpx==0.25.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
email-validator==2.1.0  # Explicitly added email-validator

# Date handling
python-dateutil==2.8.2

# Logging & Monitoring
loguru==0.7.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0